Generalization to subscribe to more than one symbol, one display
The symbols (ES1, SP500) are fixed and given, the most recent update is used
for the display update -- which works because SP500 is more current at near
"real-time" but ES1 is near 24h / 5days so after hours ES1 pops in

For randr multiple displays we currently opt to skip display 1.

Wayland has not been touched and is 'to be done'.